Player body parts now update their transforms correctly; on Android and PC. NOTE: Android seems to be flipping meshes on a single axis; changes depending on scene. Needs investigation.
# Skinned meshes will now update their transform to their nodes correctly.
# ClientData constructor is now in its implementation file.
# bodyOffsetFromHead is now sent in setup command, and converted to correct axis standard on the client.

+++ SimulCasterRenderer/src/crossplatform/SessionClient.cpp
@@ -440,7 +440,7 @@ void SessionClient::SendInput(int id,const ControllerState& controllerState)
 				packetFlags = ENET_PACKET_FLAG_UNSEQUENCED;
 			}
 		}
-		inputState.numEvents=controllerState.inputEvents.size();
+		inputState.numEvents = static_cast<uint32_t>(controllerState.inputEvents.size());
 	
 		inputBuffer.resize(sizeof(avs::InputState)+inputState.numEvents*sizeof(avs::InputEvent));
 		memcpy(inputBuffer.data(),&inputState,sizeof(avs::InputState));
